Made mode progress, Used a combination of the Factory, Tempate, and Strategy design patterns and Julias multiple dispatch feature to easily create both AC and DC models, made it easy to programm new solvers.
