+++ OOP-Refactoring/src/main.jl
@@ -1,9 +1,18 @@
 using PowerModels, JuMP, Ipopt, Gurobi
-include("power_flow_model.jl")
+include("classes.jl")
 using .MPOPF
 
 file_path = "./Cases/case5.m"
 
-ac_model = create_model_ac(file_path, Ipopt)
+ac_factory = MPOPF.ACPowerFlowModelFactory(file_path, Ipopt.Optimizer)
 
-optimize_model(ac_model)
+My_AC_model = create_model(ac_factory)
+
+optimize_model(My_AC_model)
+
+
+dc_factory = MPOPF.DCPowerFlowModelFactory(file_path, Ipopt.Optimizer)
+
+My_DC_model = create_model(dc_factory)
+
+optimize_model(My_DC_model)
